feat(padi-client): the prefix rule and the screen tail become zero-import leaves
Two pure string folds lived in `@kolu/padi/render`, behind a manifest that names
`columnify`, `kaval` and `node-pty`. Neither needs any of them, and both were
being written out again downstream because of it — "any unique prefix" has three
spellings in kolu's own tree and a fourth in a consumer, and the screen tail's
blank-viewport rule (`tail: 6` on a fresh shell returned six blank lines) had
been re-derived from the comment describing it.

`@kolu/padi-client/terminalId` and `@kolu/padi-client/screenTail`. Both files
import NOTHING, and that is the contract rather than an accident: the consumer
that asked resolves prefixes inside the package its browser bundles, so even a
type edge would put kolu's schema graph in that bundle's compile graph. Hence
`resolveTerminalId<Id extends string>` over the caller's own id type instead of
`TerminalId` — which widens nothing, `TerminalId` being a checked `Schema.String`.

`@kolu/padi/render` re-exports both, so no call site moves. The tests MOVE rather
than being copied: one left behind would pin the re-export and go quietly green
if the leaf drifted.

`shortId`/`SHORT_ID_LEN` deliberately stayed. They read as the resolver's other
half, but no consumer asked for them and kolu's own browser package holds a third
copy of the 8-char rule — publishing a one-truth entry while leaving that behind
is the second source of truth this move exists to end.
