+++ packages/padi/src/cliClient/render.test.ts
@@ -11,7 +11,6 @@ import {
   formatWatchActivityJson,
   formatWatchJson,
   formatWatchRemovalJson,
-  resolveTerminalId,
   WATCH_FEED_KINDS,
 } from "./render.ts";
 
@@ -40,51 +39,6 @@ describe("isWaitState — the whole padi-side `--until` contract", () => {
     expect(isWaitState("idle")).toBe(false);
     expect(isWaitState("")).toBe(false);
     expect(isWaitState("Awaiting")).toBe(false);
-  });
-});
-
-describe("resolveTerminalId — prefix resolution", () => {
-  const ids = [
-    "11111111-1111-1111-1111-111111111111",
-    "22222222-2222-2222-2222-222222222222",
-    "2233aaaa-0000-0000-0000-000000000000",
-  ];
-
-  it("resolves a unique prefix", () => {
-    expect(resolveTerminalId("1111", ids)).toEqual({
-      kind: "found",
-      id: ids[0],
-    });
-  });
-
-  it("prefers an exact full id over a longer id sharing its prefix", () => {
-    const withLonger = [...ids, "1111"];
-    // "1111" is an exact match AND a prefix of ids[0] — exact wins.
-    expect(resolveTerminalId("1111", withLonger)).toEqual({
-      kind: "found",
-      id: "1111",
-    });
-  });
-
-  it("is case-insensitive", () => {
-    expect(resolveTerminalId("2233AAAA", ids)).toEqual({
-      kind: "found",
-      id: ids[2],
-    });
-  });
-
-  it("reports ambiguity with the matches", () => {
-    const r = resolveTerminalId("22", ids);
-    expect(r.kind).toBe("ambiguous");
-    if (r.kind === "ambiguous") expect(r.matches).toHaveLength(2);
-  });
-
-  it("rejects an empty query as no-match (never silently the sole terminal)", () => {
-    expect(resolveTerminalId("", ids)).toEqual({ kind: "none" });
-  });
-
-  it("reports no match for an unknown prefix", () => {
-    expect(resolveTerminalId("ffff", ids)).toEqual({ kind: "none" });
   });
 });
